Revert "Otherwise trouble with pythonization in xAH"
This reverts commit b2565a750ed9c260d05b169d2c324dc42edc9ce8.


Former-commit-id: 3412d7883c691c613fa209d120885f436b553a13

+++ Reconstruction/Jet/JetRecTools/JetRecTools/JetConstituentModSequence.h
@@ -40,7 +40,7 @@ protected:
   // P-A : the actual type
   // Define as a basic integer type because Gaudi
   // doesn't support arbitrary property types
-  int m_inputType; // 
+  unsigned short m_inputType; // 
   
   
   ToolHandleArray<IJetConstituentModifier> m_modifiers;
